requirements.txt
SimpleITK==2.0.2
torch==1.12.1
monai==0.7.0
numpy==1.19.5
scikit-image==0.17.2
tqdm==4.62.2
scipy==1.5.4
pydicom==2.3.0
argparse==1.4.0
setuptools==60.10.0
pandas==1.3.5
PyYAML==6.0
gdown==4.5.1
natsort==7.0.1